pyproject.toml
[project]
name = "werewolf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "uvicorn>=0.42.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]
